Find module path at runtime
Currently, this test only passes if you run the test while being in the
source repository. Trying to run the test after installing the generated
wheel package in a different directory will fail because "./constraint" may
not exist, or worse, give a false failure since the shared libraries
might exist in the wheel package but not in the source directory.